Split the API models by subdomain
req.py and responses.py were catch-alls — a 1000-line responses.py that every
route imported — so a change to one subdomain touched files shared by all of
them.

Each subdomain now gets shvatka/api/models/<subdomain>/ with its own
requests.py and responses.py: auth, players, teams, games, waivers, files (the
/cdn game files), notifications, action_requests, search, push, admin. Models
that two or more subdomains use — Page, Items, Player, Team, Game, the identity
value objects, MergeRequest, TimelineItem — live in models/shared/; anything a
single subdomain owns stays with it.

WaiversDto and VotedPlayer were declared inline in routes/waivers.py (and
imported from there by routes/admin.py); they move to models/waivers.

routes/ is untouched apart from its import lines, so moving each route module
next to its models later is a plain file move.

Pure reorganisation: the generated OpenAPI schema is byte-identical to before.
