+++ shvatka/api/models/files/requests.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class RenameFile:
+    filename: str
