+++ shvatka/api/models/notifications/requests.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass, field
+
+
+@dataclass
+class MarkNotificationsRead:
+    ids: list[int] = field(default_factory=list)
