+++ shvatka/api/models/players/requests.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class ChangeUsername:
+    username: str
