+++ shvatka/api/models/push/responses.py
@@ -1,0 +1,7 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class PushConfigResponse:
+    enabled: bool
+    public_key: str | None
