+++ shvatka/api/routes/admin.py
@@ -9,8 +9,13 @@ from fastapi.params import Path, Query
 
 from shvatka.api.config.models.main import ApiConfig
 from shvatka.api.dependencies.auth import ApiIdentityProvider
-from shvatka.api.models import req, responses
-from shvatka.api.routes.waivers import WaiversDto
+from shvatka.api.models.admin import requests, responses
+from shvatka.api.models.shared import requests as shared_requests
+from shvatka.api.models.players import responses as players_responses
+from shvatka.api.models.shared import responses as shared
+from shvatka.api.models.games import responses as games_responses
+from shvatka.api.models.files import responses as files_responses
+from shvatka.api.models.waivers import responses as waivers_responses
 from shvatka.core.games.admin_interactors import (
     AdminUpdateGameScenarioInteractor,
     AdminUploadGameFileInteractor,
@@ -46,7 +51,7 @@ async def list_players(
     active: Annotated[bool, Query()] = True,
     archive: Annotated[bool, Query()] = False,
     can_be_author: Annotated[bool | None, Query()] = None,
-) -> responses.Items[responses.AdminPlayer]:
+) -> shared.Items[responses.AdminPlayer]:
     players = await interactor(
         identity,
         username=username,
@@ -55,7 +60,7 @@ async def list_players(
         archive=archive,
         can_be_author=can_be_author,
     )
-    return responses.Items([responses.AdminPlayer.from_core(player) for player in players])
+    return shared.Items([responses.AdminPlayer.from_core(player) for player in players])
 
 
 @inject
@@ -64,9 +69,11 @@ async def get_player(
     interactor: FromDishka[AdminGetPlayerInteractor],
     config: FromDishka[ApiConfig],
     id_: Annotated[int, Path(alias="id")],
-) -> responses.PlayerWithIdentities:
+) -> players_responses.PlayerWithIdentities:
     info = await interactor(identity, id_)
-    return responses.PlayerWithIdentities.from_core(info.player, info.email, config.superusers)
+    return players_responses.PlayerWithIdentities.from_core(
+        info.player, info.email, config.superusers
+    )
 
 
 @inject
@@ -84,15 +91,15 @@ async def change_email(
     identity: FromDishka[ApiIdentityProvider],
     interactor: FromDishka[AdminSetPlayerEmailInteractor],
     id_: Annotated[int, Path(alias="id")],
-    body: Annotated[req.AdminChangeEmail, Body()],
-) -> responses.EmailAccount:
+    body: Annotated[requests.AdminChangeEmail, Body()],
+) -> shared.EmailAccount:
     try:
         account = await interactor(
             identity=identity, player_id=id_, email=body.email, is_verified=body.verified
         )
     except exceptions.EmailAlreadyExist as e:
         raise HTTPException(status_code=409, detail="email already exists") from e
-    return responses.EmailAccount(email=account.email, is_verified=account.is_verified)
+    return shared.EmailAccount(email=account.email, is_verified=account.is_verified)
 
 
 @inject
@@ -101,15 +108,17 @@ async def change_username(
     interactor: FromDishka[AdminSetPlayerUsernameInteractor],
     config: FromDishka[ApiConfig],
     id_: Annotated[int, Path(alias="id")],
-    body: Annotated[req.AdminChangeUsername, Body()],
-) -> responses.PlayerWithIdentities:
+    body: Annotated[requests.AdminChangeUsername, Body()],
+) -> players_responses.PlayerWithIdentities:
     try:
         info = await interactor(identity=identity, player_id=id_, username=body.username)
     except exceptions.PlayerInvalidUsername as e:
         raise HTTPException(status_code=422, detail="invalid username") from e
     except exceptions.PlayerUsernameOccupied as e:
         raise HTTPException(status_code=409, detail="username already occupied") from e
-    return responses.PlayerWithIdentities.from_core(info.player, info.email, config.superusers)
+    return players_responses.PlayerWithIdentities.from_core(
+        info.player, info.email, config.superusers
+    )
 
 
 @inject
@@ -118,8 +127,8 @@ async def change_tg(
     interactor: FromDishka[AdminChangePlayerTgInteractor],
     config: FromDishka[ApiConfig],
     id_: Annotated[int, Path(alias="id")],
-    body: Annotated[req.AdminChangeTg, Body()],
-) -> responses.PlayerWithIdentities:
+    body: Annotated[requests.AdminChangeTg, Body()],
+) -> players_responses.PlayerWithIdentities:
     try:
         info = await interactor(
             identity=identity,
@@ -135,7 +144,9 @@ async def change_tg(
         raise HTTPException(
             status_code=409, detail="this telegram account is linked to another player"
         ) from e
-    return responses.PlayerWithIdentities.from_core(info.player, info.email, config.superusers)
+    return players_responses.PlayerWithIdentities.from_core(
+        info.player, info.email, config.superusers
+    )
 
 
 @inject
@@ -161,36 +172,36 @@ async def get_player_waiver_points(
     identity: FromDishka[ApiIdentityProvider],
     interactor: FromDishka[AdminGetPlayerWaiverPointsInteractor],
     id_: Annotated[int, Path(alias="id")],
-) -> responses.Items[responses.WaiverPoint]:
+) -> shared.Items[waivers_responses.WaiverPoint]:
     points = await interactor(identity, id_)
-    return responses.Items([responses.WaiverPoint.from_core(point) for point in points])
+    return shared.Items([waivers_responses.WaiverPoint.from_core(point) for point in points])
 
 
 @inject
 async def merge_players(
     identity: FromDishka[ApiIdentityProvider],
     interactor: FromDishka[AdminMergePlayersInteractor],
-    body: Annotated[req.MergePlayersRequest, Body()],
-) -> responses.Player:
+    body: Annotated[requests.MergePlayersRequest, Body()],
+) -> shared.Player:
     player = await interactor(
         identity=identity,
         primary_id=body.primary_id,
         secondary_id=body.secondary_id,
         timeline=body.core_timeline(),
     )
-    return responses.Player.from_core(player)
+    return shared.Player.from_core(player)
 
 
 @inject
 async def merge_teams(
     identity: FromDishka[ApiIdentityProvider],
     interactor: FromDishka[AdminMergeTeamsInteractor],
-    body: Annotated[req.MergeRequest, Body()],
-) -> responses.Team:
+    body: Annotated[shared_requests.MergeRequest, Body()],
+) -> shared.Team:
     team = await interactor(
         identity=identity, primary_id=body.primary_id, secondary_id=body.secondary_id
     )
-    result = responses.Team.from_core(team)
+    result = shared.Team.from_core(team)
     assert result is not None
     return result
 
@@ -200,8 +211,8 @@ async def get_waivers_by_game(
     identity: FromDishka[ApiIdentityProvider],
     interactor: FromDishka[AdminGameWaiversReaderInteractor],
     id_: Annotated[int, Path(alias="id")],
-) -> WaiversDto:
-    return WaiversDto.from_core(await interactor(identity, id_))
+) -> waivers_responses.WaiversDto:
+    return waivers_responses.WaiversDto.from_core(await interactor(identity, id_))
 
 
 @inject
@@ -211,8 +222,8 @@ async def change_game_scenario(
     dao: FromDishka[HolderDao],
     retort: FromDishka[Retort],
     id_: Annotated[int, Path(alias="id")],
-    body: Annotated[req.AdminGameScenarioEdit, Body()],
-) -> responses.FullGame:
+    body: Annotated[requests.AdminGameScenarioEdit, Body()],
+) -> games_responses.FullGame:
     game = await interactor(
         game_id=id_,
         raw_scn=body.scenario,
@@ -220,7 +231,7 @@ async def change_game_scenario(
         identity=identity,
     )
     files = await get_file_metas(game, identity, dao.game_packager)
-    return responses.FullGame.from_core(retort, game, files)
+    return games_responses.FullGame.from_core(retort, game, files)
 
 
 @inject
@@ -229,7 +240,7 @@ async def upload_game_file(
     interactor: FromDishka[AdminUploadGameFileInteractor],
     id_: Annotated[int, Path(alias="id")],
     file: Annotated[UploadFile, File()],
-) -> responses.UploadedFile:
+) -> files_responses.UploadedFile:
     content = BytesIO(await file.read())
     saved = await interactor(
         game_id=id_,
@@ -237,7 +248,7 @@ async def upload_game_file(
         original_filename=file.filename or "document",
         identity=identity,
     )
-    return responses.UploadedFile.from_core(saved)
+    return files_responses.UploadedFile.from_core(saved)
 
 
 def setup() -> APIRouter:
