+++ tests/integration/api_full/test_game_edit.py
@@ -5,7 +5,7 @@ from adaptix import Retort
 from httpx import AsyncClient
 
 from shvatka.api.dependencies.auth import AuthProperties
-from shvatka.api.models import responses
+from shvatka.api.models.games import responses
 from shvatka.core.models import dto
 from shvatka.core.models.enums import GameStatus
 from shvatka.core.services.game import create_game
