+++ tests/integration/api_full/test_notifications.py
@@ -8,7 +8,7 @@ from aiogram.types import Message
 from httpx import AsyncClient
 
 from shvatka.api.dependencies.auth import AuthProperties
-from shvatka.api.models.auth import Token
+from shvatka.api.models.auth.responses import Token
 from shvatka.core.models import dto
 from shvatka.core.utils.datetime_utils import tz_utc
 from shvatka.infrastructure.db.dao.holder import HolderDao
