+++ tests/integration/api_full/test_user.py
@@ -2,8 +2,8 @@ import pytest
 from httpx import AsyncClient
 
 from shvatka.api.dependencies.auth import AuthProperties
-from shvatka.api.models import responses
-from shvatka.api.models.auth import Token
+from shvatka.api.models.auth.responses import Token
+from shvatka.api.models.shared import responses
 from shvatka.core.models import dto
 from shvatka.infrastructure.db.dao.holder import HolderDao
